JOB: Service refactored to new location in /app/services.
