+++ src/app/services/index.ts
@@ -1,1 +1,2 @@
 export * from './instance';
+export * from './job';
