+++ src/app/theme/index.ts
@@ -1,6 +1,6 @@
 export { BaMenuService } from './services';
 export { InstanceService } from '../services';
-export { JobService } from './services';
+export { JobService } from '../services';
 export * from './theme.constants';
 export * from './theme.configProvider';
 export * from './theme.config';
