+++ src/app/theme/services/index.ts
@@ -2,4 +2,3 @@ export * from './baImageLoader';
 export * from './baMenu';
 export * from './baThemePreloader';
 export * from './baThemeSpinner';
-export * from './job-service';
